refactor: move util and helper methods to common utils
